+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Add tests to enforce behaviour of entries and index.
 * Enhance `SymbolEntryIndex` to add the ability to lookup entry by javascript name or jsni identifier.
+* Add `SymbolEntry.toString()` method that emits serialized form of symbol entry.
 
 ### [v0.01](https://github.com/realityforge/gir/tree/v0.01) (2018-03-14)
 [Full Changelog](https://github.com/realityforge/gir/compare/2edd3c5155779ee1a5b830a7056d374efc33002e...v0.01)